updated dependencies; added the ability to use dev-certs fom volga; some code cleanup and refactoring
