--- examples/sampling/cert/dev-ca.pem
@@ -1,33 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFozCCA4ugAwIBAgIUNFM95UeDqDJyOI3JWHFcvUB9wY4wDQYJKoZIhvcNAQEL
-BQAwYTELMAkGA1UEBhMCUlUxDzANBgNVBAgMBk1vc2NvdzEPMA0GA1UEBwwGTW9z
-Y293MQ4wDAYDVQQKDAVEZXZDQTEMMAoGA1UECwwDRGV2MRIwEAYDVQQDDAlEZXZS
-b290Q0EwHhcNMjUwOTIyMTQyMTMzWhcNMzUwOTIwMTQyMTMzWjBhMQswCQYDVQQG
-EwJSVTEPMA0GA1UECAwGTW9zY293MQ8wDQYDVQQHDAZNb3Njb3cxDjAMBgNVBAoM
-BURldkNBMQwwCgYDVQQLDANEZXYxEjAQBgNVBAMMCURldlJvb3RDQTCCAiIwDQYJ
-KoZIhvcNAQEBBQADggIPADCCAgoCggIBAN9yCe3zXyaL5K8EzXqllXZAnD3Ms1GC
-/d05nLmPgtDjTXavswnj00xnLLVtUmBhCBibHHKGnwx6RCWRTHpJnjePgfyVgorI
-oaVBNOaZPGtcQ344BhtZQHMnr4OlY6v/63wdvO1VqUvxBPo8tNbIpAMgg3wAAfVr
-9znOXqOz0J4aI4bsbOmZPLCrjrW+gb5u+KbB/KZPrGFODuudifnJFG06BA5zGDON
-Xt1Mc4OKAGSOlEuN95/hNXeGZ2QDRk06IRaxDL2HEev2gScICTTU/+chFXW7BGtR
-C7HwgbtP05PRh6D0fnZSky5hjiofXfTvJU5UIFTc97KfaXp9nNjoootZ/EdWjc70
-VE7iWWuAC37FtgZp7BoE3AL/cqF5gWFd9bNljzg8XweH5GMPfwAgEWHhskVTnFg3
-5i4rMVaHm/QcNRXIauhXVoWU4TlMIS6b9gy69oAZq/h5BEI/0msEWFh+1doymcZn
-ojYEA/btNJhTl0wy4W42/8VFO2A8llqmLxSGUc12PQtpb+5q7ygmFV6j3lIe396v
-1GyYYfx8jIK3d5F5G9hyCAFveXt5tn0eVzK3N/yVBpiURKA30ty8io1BETspWPU8
-HWL24ZPgtz2yMmZ5TYMlMI8i6edQ/THKldooPjuJMXuO9bFHBMNdzSw8L4OplPg9
-79miHmViUFfxAgMBAAGjUzBRMB0GA1UdDgQWBBQKABBIp/FUDshwSG0K3vM1K+yC
-5DAfBgNVHSMEGDAWgBQKABBIp/FUDshwSG0K3vM1K+yC5DAPBgNVHRMBAf8EBTAD
-AQH/MA0GCSqGSIb3DQEBCwUAA4ICAQAA0O/cGo7O6pjGmq9rJUVfXGeG2PgJaOoS
-3oroGBNWgtv5O0o09DDcyM9rBtDeT8m3oQLF9kO3DVCBa3BP7Y1rb1iDbWWHT8VT
-L78hDyuxhFBlAGySxdCXQskmKYaz8dVgiVveQYxMzwTabOxrF1lOqDLQgRePSYMI
-gzFkksUHA8ocL76XtfpLUzyYQP7sXYREKjBT6QsJesnkwyqr1C8+Y5O9CtWibmbs
-liOEAJajJTveyU9uwQstupobFpoDlglAkaj2ZWJIRq4iTbAFFV5di83eBsOepzsg
-Ph4pqta5VYVGuqXcMNxnew78+r25NeHpB97pLNgl79b9D1FqFue+ma5Rbdr6qVAE
-jvsd1hK/8BGiH39X1UkEgOAk1WHRftCpy26EG5OGfGqRhtrl9v+wgv3dhNJn8o++
-E0g14uy4ZyhwgmvVt8TbqJRgub9HLPuSqpBrNbQIcBYA9fsXWt+MAZmfcFyBuoLV
-kHcrnWaEHfBUTa7aJzHq1HW+clsDJjN4dAw9QLw5lvu7HHaZ7rJtAUNYslF3qu20
-ZCJrdZPnWpsdOnrqE61KrjXy6nBAHcfdEBk1PXEcEIjV0rMjh/uw91itxjHW5YXR
-FW+ii0GuxTyUMN/RIOsoFLTW8w7cncMxtyRpDHiTDWwMgbCUBVsT2p/0Ly/4lvbq
-O56MpIsUcg==
------END CERTIFICATE-----
