--- examples/sampling/cert/dev-server.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDLAdLW5ZzdCBs4
-vr6X9w6kfj1Pw8whTvtoGewo0RzkSK/MQhgswaG2zNb/BpC9KL4t1v027LAvSFK4
-W/AvN/MYd3APyfUvicsiblxrIUepBm7hUZFiuGzogD1NaDjzhuqN6y1rNJHEd8VE
-FfZZhtvpqFwDPTzw70ZXzpfnnnoFMRG3zKnTEsqXQEZZ5M3/To7fYwJ5vMGysQWj
-dgbpKwHH+/NzbQKUf8rUX1eViBbs2mdhrCCBETBVjLELO3zxOWf7rhXxfdVuftyB
-q13vEMT3zac/4tj/Ix5vzODcljvskNoeZMPN0wmsPU57gyzwkT6TnBQ+0kHrw2Nq
-ZVc71fGvAgMBAAECggEAXGDHj0esQn6iWIegFmnzzVGu18HK/atBxyTorMg/o/gZ
-m1LGSnOQga0j9qgHrDRCmnx2/IvnPpTkI4WevfXxVsRfB6hIvOmwuMnMn2llARru
-0PR7upce5AwbGULgi5zs1kjbAapJ3MP9zdZKpIG0R9b3i/fNZ/wSH/ngcSNR9lMG
-D/fuGuAiwUM5bhwnspiU3HKzXXwEl2dC/GFdt0XGxIcPuJHOHT5DPMhhqO/iqgYz
-JN7j7GP0Chabr9f5P0hxajFhaTvKvvrT9n16hZoISLgkRf9EcxZ2x2iWTovCOrcB
-sgIMXs+XU9m3CSPphsMyjzqb5MiSEumMwbdGUKhYnQKBgQD6Xm05VSy7zk80EmIQ
-KvN6r4wEM1//ApXvLW3o+LFkFZP2VETPHselojxmoWcFYacTkNO4BT0ft3htYE9r
-6Aby/thPId0wSZBVa490/HnqX67pzjU0OGVb8yvj6zMtT4ICuSZNCHf9tl0h4wH0
-w18/kCYUtAGQQ+JMORsqHclKGwKBgQDPkrKY9Giz43ZnEyExdbRwn+9fr5ZoEf8B
-d6WUOALLUrzVA/vYY/W3D6hHZbWoqNAGWzrNFlOUD7+0PYHtaNx48IBeNDXR9aq8
-LSqeZeqffqJrp0F4pqD4Pvl79gauVHXECt5GjhvLHdJP04IX9XYBYQsovzi55qZL
-5qwH2rJv/QKBgB4Ts+lcSxhwq2ElCDTLBB7r8dUbWt/mrp5iBb8kCSNt2C2SaAW9
-u9wxBjKKFzc1SP33uwdzUal6v/Pbyw66fAM7DD9NYSYI4FbyqfQy1G6MehXKWOfi
-EoMeTlp+CdLap1KubWR5rPiErOGCRNzjj5CyP3DjQwDStt/TSGkh1OmbAoGAQww+
-ljt0So7xou/B7Aj6JNr0ZjMAYVNWzHGKUWsaM/Tgu+nSgp4fM2diUtWerQECWNtF
-DhPlIVFI+f5YJoiVlpcQz90LeWtiH/By8ME43RX5fs6Czmb8XdNxoMHifJMZix57
-l8y1W5GhalmGjxihamafRSUnPJ0tIB09ejgBJT0CgYEArGntNLPK3+v1uJk0CdRf
-rC0smCDY3FWi2sPwUVSVxpU2IC3bE0MCQ1WwAP2XuqpigGnqvHdJQPBKEIk8HAvw
-NqKzsRrDXQxoDgAIZAsxiahkC7IErJvdj0AOj/zfkvWvhiqENK/KEyh5lGmGbw9+
-08o9usn/rwBm0USrYlAW15Q=
------END PRIVATE KEY-----
