--- examples/sampling/cert/dev-server.pem
@@ -1,27 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIEljCCAn6gAwIBAgIUVvJwY9q7dnmxwi7j2WpnkjnHHV8wDQYJKoZIhvcNAQEL
-BQAwYTELMAkGA1UEBhMCUlUxDzANBgNVBAgMBk1vc2NvdzEPMA0GA1UEBwwGTW9z
-Y293MQ4wDAYDVQQKDAVEZXZDQTEMMAoGA1UECwwDRGV2MRIwEAYDVQQDDAlEZXZS
-b290Q0EwHhcNMjUwOTIyMTQyMjI2WhcNMjYwOTIyMTQyMjI2WjBlMQswCQYDVQQG
-EwJSVTEPMA0GA1UECAwGTW9zY293MQ8wDQYDVQQHDAZNb3Njb3cxEjAQBgNVBAoM
-CURldlNlcnZlcjEMMAoGA1UECwwDRGV2MRIwEAYDVQQDDAlsb2NhbGhvc3QwggEi
-MA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDLAdLW5ZzdCBs4vr6X9w6kfj1P
-w8whTvtoGewo0RzkSK/MQhgswaG2zNb/BpC9KL4t1v027LAvSFK4W/AvN/MYd3AP
-yfUvicsiblxrIUepBm7hUZFiuGzogD1NaDjzhuqN6y1rNJHEd8VEFfZZhtvpqFwD
-PTzw70ZXzpfnnnoFMRG3zKnTEsqXQEZZ5M3/To7fYwJ5vMGysQWjdgbpKwHH+/Nz
-bQKUf8rUX1eViBbs2mdhrCCBETBVjLELO3zxOWf7rhXxfdVuftyBq13vEMT3zac/
-4tj/Ix5vzODcljvskNoeZMPN0wmsPU57gyzwkT6TnBQ+0kHrw2NqZVc71fGvAgMB
-AAGjQjBAMB0GA1UdDgQWBBRSPG1jSgs/0tML3cdVW8WN1KXJpTAfBgNVHSMEGDAW
-gBQKABBIp/FUDshwSG0K3vM1K+yC5DANBgkqhkiG9w0BAQsFAAOCAgEAtu5vZzPW
-hr4KF2hosX1amj2uP3JSp6Cr9Vf9F7qQz7yh046oiLN+vkj9iJWoudZ+WWr8wNuk
-PnSUsDVHBm3Qz11rlDzavyVo288mhZeWIcsOeOBWkaADQNoiIC7Fh/CIUkTBxsuB
-3NcR/qgxlCpqvErXfKQbw1YAohXC51bcXGdCB424GHPqOUZrlzLyB4xiwrRl0/aH
-vdNnZjuUV8UXrpFr3qlPeb9/T5bCRXKaaNn7wtlogIDL4682BcTA7rSh9QCOh2Lq
-kuaiVDVISUYnj/uCin+DBddMUQ/13pSxfO6ikg2hxpy8zMwYEwK2QVxNfJ2g2/oT
-U8A8B2Ih7mWZKvfaj29JMOhbt/7DNdPzUnkVVP04FS41AG+Es+nJnfTJgKyC6h9Y
-dBAkeZPO+3VgB8a7bZCpkg5/kuxfntFZToFdWcW0cpJ55dG+Zka5l6o2/uGLvNQE
-2EGVnhpFKZyZFLBkAg3Wo7PV9UKHTkpitCYrTvdNRlQHpGcr/LVvD5br+bcxZHe+
-yg5nvzgcR/GqZC7x3rZ3cq8qvxylNDUpnHtZrSR7Va6DtUWcGA+lW6nTZZZ7kg9G
-ocNKyiKdPdDzDl2nkeQbzHdhT1LXcBBilvLj+RmBaaSw088AxG8zAOLsIeZGHs2L
-yJ+RCavum+Na2fL4XfYG7Ow2ZUsf0qOFSPU=
------END CERTIFICATE-----
