+++ neva/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "neva"
-version = "0.1.7"
+version = "0.1.8"
 edition = "2021"
 rust-version = "1.85.0"
 authors = ["Roman Emreis <roman.emreis@outlook.com>"]
@@ -20,10 +20,10 @@ dashmap = "6.1.0"
 futures-util = { version = "0.3.31", default-features = false, features = ["alloc"] }
 inventory = { version = "0.3.21", optional = true }
 jsonschema = { version = "0.33.0", optional = true }
-memchr = "2.7.5"
+memchr = "2.7.6"
 once_cell = { version = "1.21.3", features = ["std"], optional = true }
 reqwest = { version = "0.12.23", features = ["stream", "json"], optional = true }
-serde = { version = "1.0.226", features = ["derive"] }
+serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
 sse-stream = { version = "0.2.1", optional = true }
 schemars = { version = "1.0.4" }
@@ -32,9 +32,11 @@ tokio-stream = { version = "0.1.17", optional = true }
 tokio-util = "0.7.16"
 tracing = { version = "0.1.41", optional = true }
 tracing-subscriber = { version = "0.3.20", features = ["fmt", "json"], optional = true }
-volga = { version = "0.6.4", features = ["di", "jwt-auth-full"], optional = true }
+volga = { version = "0.6.6", features = ["di", "jwt-auth-full"], optional = true }
 uuid = { version = "1.18.1", features = ["v4", "serde"] }
-neva_macros = { path = "../neva_macros", version = "0.1.7", optional = true }
+
+# neva
+neva_macros = { path = "../neva_macros", version = "0.1.8", optional = true }
 
 [target.'cfg(target_os = "windows")'.dependencies]
 windows = { version = "0.62.0", features = ["Win32_Foundation", "Win32_System_JobObjects", "Win32_System_Threading", "Win32_Security", "Win32_System_Diagnostics", "Win32_System_Diagnostics_ToolHelp"], optional = true }
@@ -59,7 +61,7 @@ server-macros = ["server", "macros", "neva_macros?/server"]
 client = ["dep:windows", "dep:nix", "dep:jsonschema", "tokio/process", "tokio/signal"]
 server = ["tokio/signal"]
 client-tls = ["reqwest?/rustls-tls"]
-server-tls = ["volga?/tls"]
+server-tls = ["volga?/tls", "volga?/dev-cert"]
 tracing = ["dep:tracing", "dep:tracing-subscriber", "dep:once_cell", "volga?/tracing"]
 
 [package.metadata.docs.rs]
