+++ neva/src/lib.rs
@@ -109,7 +109,7 @@ pub mod prelude {
     #[cfg(feature = "http-server")]
     pub use crate::transport::HttpServer;
     #[cfg(all(feature = "http-server", feature = "server-tls"))]
-    pub use crate::transport::http::TlsConfig;
+    pub use crate::transport::http::{TlsConfig, DevCertMode};
     
     #[cfg(feature = "server")]
     pub use crate::app::{App, context::Context, options};
