+++ neva_macros/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "neva_macros"
-version = "0.1.7"
+version = "0.1.8"
 edition = "2021"
 rust-version = "1.85.0"
 authors = ["Roman Emreis <roman.emreis@outlook.com>"]
@@ -16,7 +16,7 @@ keywords = ["modelcontextprotocol", "mcp", "sdk", "ai", "framework"]
 proc-macro = true
 
 [dependencies]
-quote = "1.0.40"
+quote = "1.0.41"
 syn = { version = "2.0.106", features = ["full"] }
 proc-macro2 = "1.0.101"
