+++ neva_macros/README.md
@@ -1,7 +1,7 @@
 # Neva Macros
 Macros crate for Neva MCP SDK
 
-[![latest](https://img.shields.io/badge/latest-0.1.6-d8eb34)](https://crates.io/crates/neva)
+[![latest](https://img.shields.io/badge/latest-0.1.8-d8eb34)](https://crates.io/crates/neva)
 [![latest](https://img.shields.io/badge/rustc-1.85+-964B00)](https://crates.io/crates/neva)
 [![License: MIT](https://img.shields.io/badge/License-MIT-624bd1.svg)](https://github.com/RomanEmreis/neva/blob/main/LICENSE)
 [![CI](https://github.com/RomanEmreis/neva/actions/workflows/rust.yml/badge.svg)](https://github.com/RomanEmreis/neva/actions/workflows/rust.yml)
